% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixtureFit.R
\docType{methods}
\name{BayesianMixtureFit}
\alias{BayesianMixtureFit}
\alias{BayesianMixtureFit,EventData,PriorAndInitsMixture-method}
\title{BayesianMixtureFit of the Event Data}
\usage{
BayesianMixtureFit(object, prior.init, ...)

\S4method{BayesianMixtureFit}{EventData,PriorAndInitsMixture}(object,
  prior.init, dist = "weibull", N.samples = 2000, N.burnin = 1000,
  N.chains = 1, N.thin = 1, seed = NULL, parallel = FALSE,
  N.proc = NULL, ...)
}
\arguments{
\item{object}{The EventData object}

\item{prior.init}{Initial values to parameters and prior values, see 
\code{PriorAndInitValues}}

\item{...}{Additional arguments to be passed to the method}

\item{dist}{Distribution (only Weibull available for mixture). Can "force" into
exponential by specifying a tight boundary around 1 for the shape parameter.}

\item{N.samples}{The number of samples to take (2000). This will also determine
the number of simulations used in the prediction step.}

\item{N.burnin}{Number of burnin iterations before sampling starts (1000). Before
these 1000 adaptive iterations are tun to enhance the sampling efficiency.}

\item{N.chains}{Number of chins to use. IF running on a parallel machine, one may
match this with the number of processors (N.proc).}

\item{N.thin}{Thining interval to use. May reduce auto-correlation.}

\item{seed}{Random seed, integer (NULL)}

\item{parallel}{Use more than one processor (FALSE)}

\item{N.proc}{Number of processors to use if running on a cluster.}
}
\value{
\code{EventData} object with the data censored at the appropriate point
}
\description{
Creates an EventModel based on the fitting a Bayesian MCMC mixture model using
run.jags and jags.

BayesianMixtureFit the Event Data
}

