% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkValidTimeCut}
\alias{checkValidTimeCut}
\title{Checks that the change point is ok}
\usage{
checkValidTimeCut(x, y)
}
\arguments{
\item{x}{EventModel object}

\item{y}{EventModelExtended object}
}
\description{
Checks that the change point is ok
}

