% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelBayesian.R
\docType{methods}
\name{densityPlot}
\alias{densityPlot}
\alias{densityPlot,EventModelBayesian-method}
\title{Plots the densities for each parameter of the fitted object}
\usage{
densityPlot(object, ...)

\S4method{densityPlot}{EventModelBayesian}(object)
}
\arguments{
\item{object}{EventModelBayesian object}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Plots the densities for each parameter of the fitted object
}

