% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagTimeEstimate.R
\docType{methods}
\name{estimateLagTimeG}
\alias{estimateLagTime}
\alias{estimateLagTime,EventData-method}
\alias{estimateLagTimeG}
\title{GMethod to estimate lag times}
\usage{
estimateLagTime(object, t.start, t.stop, dt, dist = "exponential",
  criterion = "AIC")

\S4method{estimateLagTime}{EventData}(object, t.start, t.stop, dt,
  dist = "exponential", criterion = "AIC")
}
\arguments{
\item{object}{EventData object}

\item{t.start}{Start time for search interval [Integer, months]}

\item{t.stop}{end time for search interval [Integer, months]}

\item{dt}{Time increment for search [Integer, months]}

\item{dist}{Survival distribution, see survival package ("exponential")}

\item{criterion}{Estimation criterion for the point estimate, AIC or BIC (AIC).}
}
\description{
Methods to estimate lag times

Methods to estimate lag times using survival package
}

