% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{EventModelBayesian-class}
\alias{EventModelBayesian-class}
\title{An S4 class containing a fitted survival model obtained from JAGS
of an Eventdata object}
\description{
An S4 class containing a fitted survival model obtained from JAGS
of an Eventdata object
}
\section{Slots}{

\describe{
\item{\code{model}}{A "runjags" object}

\item{\code{event.data}}{An EventData object used to fit a Weibull survial model}

\item{\code{shape.median}}{Shape median of sampled distribution.}

\item{\code{scale.1.median}}{Median of the first of the scale parameters of sampled distribution.}

\item{\code{scale.2.median}}{Median of rthe second of the scale parameters of sampled distribution.}

\item{\code{pi.median}}{Median of the mixture coefficient of the sampled distribution.}

\item{\code{mcmc.object}}{The MCMC object containing the sampled distributions.}

\item{\code{seed}}{Seed used stored for reproducibility.}

\item{\code{jags.model}}{The jags model used as a character string.}
}}

