% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelExtended.R
\docType{class}
\name{EventModelExtended-class}
\alias{EventModelExtended-class}
\title{An S4 class containing a fitted survival model
of an Eventdata object}
\description{
An S4 class containing a fitted survival model
of an Eventdata object
}
\section{Slots}{

\describe{
\item{\code{model}}{An S3 "weibreg" object}

\item{\code{event.data}}{An EventData object used to fit a Weibull survial model}

\item{\code{time.cut}}{The change point}

\item{\code{simParams}}{The \code{FromDataSimParam} object which contains the information
needed to generate the survial times}
}}

