% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelSubgroup.R
\docType{class}
\name{EventModelSubgroup-class}
\alias{EventModelSubgroup-class}
\title{An S4 class containing a fitted survival model
of an Eventdata object}
\description{
An S4 class containing a fitted survival model
of an Eventdata object
}
\section{Slots}{

\describe{
\item{\code{model.1}}{An S3 "survreg" object}

\item{\code{model.2}}{An S3 "survreg" object}

\item{\code{s1}}{Name of subgroup 1}

\item{\code{s2}}{Name of subgroup 2}

\item{\code{event.data}}{An EventData object used to fit a Weibull survial model}

\item{\code{simParams}}{The \code{FromDataSimParam} object which contains the information
needed to generate the survial times}
}}

