% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelExtended.R
\name{getSegment}
\alias{getSegment}
\title{Help-function to get a segment for plotting}
\usage{
getSegment(x.start, x.end, y.start, npts, shape, scale)
}
\arguments{
\item{x.start}{x start coordinate}

\item{x.end}{x end coordinate}

\item{y.start}{y start coordinate}

\item{npts}{number of points}

\item{shape}{Weibull shape parameter}

\item{scale}{Weibull scale parameter}
}
\description{
Help-function to get a segment for plotting
}

