% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{KMRightCensor,EventData-method}
\alias{KMRightCensor,EventData-method}
\title{KMRightCensor the Event Data}
\usage{
\S4method{KMRightCensor}{EventData}(object, time.cut)
}
\arguments{
\item{object}{The EventData object}

\item{time.cut}{The time in days when to cut the data}
}
\description{
KMRightCensor the Event Data
}

