% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lagTimeEstimate.R
\docType{class}
\name{LagTimeEstimate-class}
\alias{LagTimeEstimate-class}
\title{An S4 class containing an LagTimeEstimate Object}
\description{
An S4 class containing an LagTimeEstimate Object
}
\section{Slots}{

\describe{
\item{\code{event.data}}{An EventData object used to fit a Weibull survial model}

\item{\code{XIC}}{vector of AIC or BIC values of fits at different times.}

\item{\code{estimate.lag.time}}{Estimate of lag time}

\item{\code{criterion}}{AIC or BIC (AIC)}
}}

