% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leftTruncateFit.R
\docType{methods}
\name{LeftFit,EventData,numeric-method}
\alias{LeftFit,EventData,numeric-method}
\title{LeftFit the Event Data}
\usage{
\S4method{LeftFit}{EventData,numeric}(object, time.cut, dist = "weibull", ...)
}
\arguments{
\item{object}{The EventData object}

\item{time.cut}{The time in days when to cut the data}

\item{dist}{Distribution (weibull)}

\item{...}{Additional arguments to be passed to the method}
}
\description{
LeftFit the Event Data
}

