% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelSubgroup.R
\docType{methods}
\name{params}
\alias{params}
\alias{params,EventModelSubgroup-method}
\title{Plots the traces for each parameter of the fitted object}
\usage{
params(object, ...)

\S4method{params}{EventModelSubgroup}(object)
}
\arguments{
\item{object}{EventModelSubgroup object}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Plots the traces for each parameter of the fitted object
}

