% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eventModelExtended.R, R/lagTimeEstimate.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,EventModel,EventModelExtended-method}
\alias{plot,EventModelExtended,missing-method}
\alias{plot,LagTimeEstimate,missing-method}
\title{Plot method for LagTimeEstimate objects}
\usage{
\S4method{plot}{EventModelExtended,missing}(x, units = "Days",
  xlab = paste("Time in study [", units, "]", sep = ""), ylab = "",
  main = "", ylim = NULL, xlim = NULL, ...)

\S4method{plot}{EventModel,EventModelExtended}(x, y, units = "Days",
  xlab = paste("Time in study [", units, "]", sep = ""), ylab = "",
  main = "", ylim = NULL, xlim = NULL, ...)

\S4method{plot}{LagTimeEstimate,missing}(x, smoothing = TRUE, ...)
}
\arguments{
\item{units}{Scale for the x-axis. "Days", "Months" or "Years"}

\item{smoothing}{If true will add a smoothed line to the curve and use
that for when retreiving the optimal time value.}

\item{units}{Scale for the x-axis. "Days", "Months" or "Years"}
}
\description{
Plot method for LagTimeEstimate objects
}

