% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelExtended.R
\name{plotSegment}
\alias{plotSegment}
\title{Help-function that plots a segment}
\usage{
plotSegment(seg, col, xscale)
}
\arguments{
\item{seg}{segment to plot}

\item{col}{colour}

\item{xscale}{scale for x-axis}
}
\description{
Help-function that plots a segment
}

