% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixtureFit.R
\docType{class}
\name{PriorAndInitsMixture-class}
\alias{PriorAndInitsMixture-class}
\title{An S4 class containing a fitted survival model obtained from JAGS
of an Eventdata object}
\description{
An S4 class containing a fitted survival model obtained from JAGS
of an Eventdata object
}
\section{Slots}{

\describe{
\item{\code{shape.init}}{Initial value for shape parameter.}

\item{\code{shape.min}}{The shape parameter prior is uniformly distributed between 
[shape.min, shape.max].}

\item{\code{shape.max}}{The shape parameter prior is uniformly distributed between 
[shape.min, shape.max].}

\item{\code{ctrl.median}}{Control median [months]}

\item{\code{hr}}{Hazard ratio}

\item{\code{mixture.min}}{The minimum mixture coefficient, e.g. randomization 
balance (boundary of uniform distribution), prior is uniformly distributed between 
[mixture.min, mixture.max]}

\item{\code{mixture.max}}{The maximum mixture coefficient, e.g. randomization 
balance (boundary of uniform distribution), prior is uniformly distributed between 
[mixture.min, mixture.max]}

\item{\code{scale.sd.min}}{The minimum standard deviation of the log(scale) (0.5), 
prior is uniformly distributed between [scale.sd.min, scale.sd.max]}

\item{\code{scale.sd.max}}{The maximum standard deviation of the log(scale) (2), 
prior is uniformly distributed between [scale.sd.min, scale.sd.max]}
}}

