% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesianMixtureFit.R
\name{PriorAndInitValues}
\alias{PriorAndInitValues}
\title{PriorAndInitsMixture constructor}
\usage{
PriorAndInitValues(shape.init, shape.min, shape.max, ctrl.median, hr,
  mixture.min, mixture.max, scale.sd.min = 0.9, scale.sd.max = 1.1)
}
\arguments{
\item{shape.init}{Initial value of the shape parameter.}

\item{shape.min}{The shape parameter prior is uniformly distributed between 
[shape.min, shape.max].}

\item{shape.max}{The shape parameter prior is uniformly distributed between 
[shape.min, shape.max].}

\item{ctrl.median}{Control median [months]}

\item{hr}{Hazard ratio}

\item{mixture.min}{The minimum mixture coefficient, e.g. randomization 
balance (boundary of uniform distribution), prior is uniformly distributed between 
[mixture.min, mixture.max]}

\item{mixture.max}{The maximum mixture coefficient, e.g. randomization 
balance (boundary of uniform distribution), prior is uniformly distributed between 
[mixture.min, mixture.max]}

\item{scale.sd.min}{The minimum standard deviation of the log(scale) (0.5), 
prior is uniformly distributed between [scale.sd.min, scale.sd.max]}

\item{scale.sd.max}{The maximum standard deviation of the log(scale) (2), 
prior is uniformly distributed between [scale.sd.min, scale.sd.max]}
}
\description{
Creates a PriorAndInitsMixture object. These are guesses of the original 
parameters based on (protocol) assumptions.
}

