% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulatePW.R
\name{rcpweibull}
\alias{rcpweibull}
\title{Conditional function for piece wise model}
\usage{
rcpweibull(t.conditional, params, HR)
}
\arguments{
\item{params}{paramter object with study settings}

\item{hr}{HR}
}
\description{
Conditional function for piece wise model
}

