% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate.Mixture}
\alias{simulate.Mixture}
\title{Only used internally}
\usage{
simulate.Mixture(data, params, accrualGenerator = NULL, Naccrual = 0,
  limit = 0.05, longlagsettings = NULL, dropout = NULL)
}
\arguments{
\item{data}{EventData object}

\item{params}{Parameters from jags-sampling}

\item{accrualGenerator}{accrualGenerator object}

\item{Naccrual}{Number of subjects to recruit}

\item{limit}{Used to set CI limit}

\item{longlagsettings}{LonLagSettings object}

\item{dropout}{Dropout object}
}
\description{
Only used internally
}

