% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSub.R
\name{simulate.Subgroup}
\alias{simulate.Subgroup}
\title{Only used internally}
\usage{
simulate.Subgroup(data, s1, s2, SimParams.s1, SimParams.s2, Nsim, seed,
  accrualGenerator = NULL, Naccrual = 0, limit = 0.05,
  longlagsettings = NULL, dropout = NULL)
}
\arguments{
\item{data}{EventData object}

\item{s1}{Name (label) of subgroup 1}

\item{s2}{Name (label) of subgroup 2}

\item{SimParams.s1}{Parameters for fitted for first subgroup}

\item{SimParams.s2}{Parameters for fitted for second subgroup}

\item{Nsim}{Number of simulations}

\item{seed}{Random seed}

\item{accrualGenerator}{accrualGenerator object}

\item{Naccrual}{Number of subjects to recruit}

\item{limit}{Used to set CI limit}

\item{longlagsettings}{LonLagSettings object}

\item{dropout}{Dropout object}
}
\description{
Only used internally
}

