% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateMixW.R
\docType{methods}
\name{SimulateMixW}
\alias{SimulateMixW}
\alias{SimulateMixW,EventModelBayesian-method}
\title{Simulate method for use when predicting from piecewise Weibull}
\usage{
SimulateMixW(object, ...)

\S4method{SimulateMixW}{EventModelBayesian}(object, ...)
}
\arguments{
\item{object}{The fitted mixture model}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Simulate method for use when predicting from piecewise Weibull
}

