% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulatePW.R
\docType{methods}
\name{simulatePW}
\alias{simulatePW}
\alias{simulatePW,EventModel,EventModelExtended-method}
\title{Simulate method for use when predicting from piecewise Weibull}
\usage{
simulatePW(object, object2, ...)

\S4method{simulatePW}{EventModel,EventModelExtended}(object, object2, ...)
}
\arguments{
\item{object}{The model fitted on Right censored data}

\item{object2}{The model fittend on Left truncated/Right censored data}
}
\description{
Simulate method for use when predicting from piecewise Weibull
}

