% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateSub.R
\docType{methods}
\name{simulateSub}
\alias{simulateSub}
\alias{simulateSub,EventModelSubgroup-method}
\title{Simulate method for use when predicting from piecewise Weibull}
\usage{
simulateSub(object, ...)

\S4method{simulateSub}{EventModelSubgroup}(object, ...)
}
\arguments{
\item{object}{The fitted mixture model}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Simulate method for use when predicting from piecewise Weibull
}

