% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{splineResolution}
\alias{splineResolution}
\title{Sets the smoothing parameter, usally in (0,1].}
\usage{
splineResolution()
}
\value{
The spline resolution
}
\description{
Sets the smoothing parameter, usally in (0,1].
}

