% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subgroupFit.R
\docType{methods}
\name{SubgroupFit,EventData-method}
\alias{SubgroupFit,EventData-method}
\title{SubgroupFit the Event Data}
\usage{
\S4method{SubgroupFit}{EventData}(object, dist.1 = "weibull",
  dist.2 = "weibull", ...)
}
\arguments{
\item{object}{The EventData object}

\item{...}{Additional arguments to be passed to the method}

\item{time.cut}{The time in days when to cut the data}

\item{dist}{Distribution (weibull)}
}
\description{
SubgroupFit the Event Data
}

