% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getJagsModel}
\alias{getJagsModel}
\title{Function defining jags model}
\usage{
getJagsModel(prior.init, ctrl.rate, exp.rate)
}
\arguments{
\item{prior.init}{Prior initialization parameters}

\item{ctrl.rate}{Rate in subgroup 1}

\item{exp.rate}{Rate in subgroup 2}
}
\description{
Function defining jags model
}

