% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LeftFit}
\alias{LeftFit}
\title{LeftFit the Event Data}
\usage{
LeftFit(object, time.cut, ...)
}
\arguments{
\item{object}{The EventData object}

\item{time.cut}{The time in days when to cut the data}

\item{...}{Additional arguments to be passed to the method}
}
\value{
\code{EventData} object with the data censored at the appropriate point
}
\description{
Creates an EventModel based on the left truncated data
}

