% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leftTruncateFit.R
\docType{methods}
\name{LeftFit}
\alias{LeftFit}
\title{LeftFit the Event Data}
\usage{
LeftFit(object, time.cut, ...)

\S4method{LeftFit}{EventData,numeric}(object, time.cut, dist = "weibull")
}
\arguments{
\item{object}{The EventData object}

\item{time.cut}{The time in days when to cut the data}

\item{dist}{Distribution (weibull)}
}
\value{
\code{EventData} object with the data censored at the appropriate point
}
\description{
Creates an EventModel based on the left truncated data
}

