% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulatePW.R
\name{rcpweibull}
\alias{rcpweibull}
\title{Conditional function for piece wise model}
\usage{
rcpweibull(t.conditional, params, HR)
}
\arguments{
\item{t.conditional}{time vector to condition on}

\item{params}{paramter object with study settings}

\item{HR}{HR}
}
\description{
Conditional function for piece wise model
}

