% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{show,EventModelBayesian-method}
\alias{show,EventModelExtended-method}
\title{Displays information about EventModelBayesian objects}
\usage{
\S4method{show}{EventModelBayesian}(object)

\S4method{show}{EventModelExtended}(object)
}
\arguments{
\item{object}{EventModelExtended object}
}
\description{
Displays information about EventModelBayesian objects

Displays information about EventModelExtended objects
}

