% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelBayesian.R, R/eventModelExtended.R, R/eventModelSubgroup.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,EventModelBayesian-method}
\alias{show,EventModelExtended-method}
\alias{show,EventModelSubgroup-method}
\title{Displays information about EventModelBayesian objects}
\usage{
\S4method{show}{EventModelBayesian}(object)

\S4method{show}{EventModelExtended}(object)

\S4method{show}{EventModelSubgroup}(object)
}
\arguments{
\item{object}{EventModelExtended object}

\item{object}{EventModelSubgroup object}
}
\description{
Displays information about EventModelBayesian objects

Displays information about EventModelExtended objects

Displays information about EventModelSubgroup objects
}

