% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eventModelExtended.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,EventModelExtended-method}
\title{Displays information about EventModelExtended objects}
\usage{
\S4method{show}{EventModelExtended}(object)
}
\arguments{
\item{object}{EventModelExtended object}
}
\description{
Displays information about EventModelExtended objects
}

