% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulatePW.R
\name{simulate.Internal}
\alias{simulate.Internal}
\title{Only used internally}
\usage{
simulate.Internal(data, SimParamsRgt, SimParamsLft, time.cut,
  accrualGenerator = NULL, Naccrual = 0, Nsim = 10000, seed = NULL,
  limit = 0.05, longlagsettings = NULL, dropout = NULL)
}
\description{
Only used internally
}

