% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simulate.Internal}
\alias{simulate.Internal}
\title{Only used internally}
\usage{
\method{simulate}{Internal}(data, SimParamsRgt, SimParamsList, timeCutList,
  accrualGenerator = NULL, Naccrual = 0, Nsim = 10000, seed = NULL,
  limit = 0.05, longlagsettings = NULL, dropout = NULL)
}
\arguments{
\item{data}{EventData object}

\item{SimParamsRgt}{Parameters for fitted KM right censored model}

\item{SimParamsList}{List of parameters for Fitted left truncated model}

\item{timeCutList}{List of change points, should equal length of SimParamsList}

\item{accrualGenerator}{accrualGenerator object}

\item{Naccrual}{Number of subjects to recruit}

\item{Nsim}{Number of simulations}

\item{seed}{Random seed}

\item{limit}{Used to set CI limit}

\item{longlagsettings}{LonLagSettings object}

\item{dropout}{Dropout object}
}
\description{
Only used internally
}

