% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePW.R
\docType{methods}
\name{simulatePW}
\alias{simulatePW}
\alias{simulatePW,EventModel,EventModelExtended-method}
\alias{simulatePW,EventModel,list-method}
\title{Simulate method for use when predicting from piecewise Weibull}
\usage{
simulatePW(object, leftobjects, ...)

\S4method{simulatePW}{EventModel,EventModelExtended}(object, leftobjects, ...)

\S4method{simulatePW}{EventModel,list}(object, leftobjects, ...)
}
\arguments{
\item{object}{The model fitted on Right censored data}

\item{leftobjects}{The model fitted on Left truncated/Right censored data. Either a single
or a list of many EventModelExtended objects.}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Simulate method for use when predicting from piecewise Weibull
}

