% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventModelBayesian.R
\docType{methods}
\name{tracePlot}
\alias{tracePlot}
\alias{tracePlot,EventModelBayesian-method}
\title{Plots the traces for each parameter of the fitted object}
\usage{
tracePlot(object, ...)

\S4method{tracePlot}{EventModelBayesian}(object)
}
\arguments{
\item{object}{EventModelBayesian object}

\item{...}{Additional arguments to be passed to the method}
}
\description{
Plots the traces for each parameter of the fitted object
}

